Near Completion

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -18,8 +19,35 @@ class UserController extends Controller
         ]);
     }
 
+    public function store(Request $request)
+    {
+        $data = $request->validate([
+            'name' => ['required'],
+            'email' => ['required', 'email', 'unique:users'],
+            'password' => ['required', 'min:8'],
+            'role' => ['required', 'in:admin,issuer,employer'],
+        ]);
+
+        $user = User::create([
+            'name' => $data['name'],
+            'email' => $data['email'],
+            'password' => Hash::make($data['password']),
+        ]);
+
+        $user->assignRole($data['role']);
+
+        return back()->with('flash', [
+            'type' => 'success',
+            'message' => 'User created successfully',
+        ]);
+    }
+
     public function destroy(Request $request, User $user)
     {
+        if ($request->user()->hasRole('admin') && $user->hasRole('admin')) {
+            abort(403);
+        }
+
         if ($request->user()->hasRole('admin')) {
             $user->delete();
             return redirect()->back()->with('success', 'You have successfully deleted the user');
